+++ example/xml_example.dart
@@ -1,5 +1,5 @@
 import 'package:processing_tree/processing_tree.dart';
-import 'package:processing_tree/tree_builder.dart';
+import 'package:processing_tree/src/tree_builder.dart';
 
 class CalculatorCtx {
   double inValue = 0;