+++ lib/src/tree_builder.dart
@@ -1,5 +1,3 @@
-library processing_tree;
-
 import 'package:xml/xml.dart';
 
 import 'tree_processor.dart';
@@ -24,7 +22,7 @@ class TreeBuilder {
   /// If [parent] is null then root node is add, this can be done only once or
   /// [TreeBuilderException] will be thrown. Each next call to this method
   /// should provide [parent] object for newly added node. As a parent to
-  /// subnode object returned from previous call should be used:
+  /// subNode object returned from previous call should be used:
   ///
   /// ```dart
   ///   var root = builder.add(null, (c,d) => {return Action.proceed}, null);
@@ -238,19 +236,129 @@ abstract class DelegateProvider {
   dynamic delegateData(String delegateName, Map<String, dynamic> rawData);
 }
 
-class _ParsedItem {
+/// Tells builder if given node should be considered as a value or algorithm
+/// element
+enum ParsedItemType {
+  /// This is node which represents constant value. While parsing call to
+  /// delegate will be performed, returned value will be passed to parent node
+  /// [data] collection. Item marked with this type will not create node in
+  /// processing tree!
+  ///
+  /// *Contract #1:* delegate of constValue takes KeyValue object as a context
+  /// and can change name/value fields as needed.
+  ///
+  /// *Contract #2:* Parent data is not null and it conforms to
+  /// operator [](String name)=dynamic. To store result of delegate call.
+  ///
+  /// *Contract #3:* Result value of delegate different then [Action.proceed] is
+  /// considered as error and TreBuildException is thrown.
+  ///
+  /// *Contract #4:* ConstValue might have any nested children, but all of them
+  /// must be also a constValue.
+  constValue,
+
+  /// This is node which is part of algorithm, it should be represented as
+  /// a delegate and data in Processing Tree
+  owner
+}
+
+class _WrappedDelegateProvider extends BuildCoordinator {
+  final DelegateProvider _provider;
+
+  _WrappedDelegateProvider(this._provider);
+
+  @override
+  void step(BuildAction action, ParsedItem item) {
+    // Nothing
+  }
+
+  @override
+  ParsedItem requestData(BuildPhaseState state) {
+    final delegate = _provider.delegate(state.delegateName)!;
+    final data = _provider.delegateData(state.delegateName, state.data);
+    return ParsedItem.from(state, delegate, data, ParsedItemType.owner);
+  }
+}
+
+/// Informs [BuildCoordinator] about new step in parse process
+enum BuildAction {
+  /// New item which is owner type is about to be processed. This action is
+  /// before any other call to coordinator
+  newItem,
+  
+  /// Item was just finalised, no more actions for it will be done.
+  finaliseItem,
+
+  /// Start processing of children of current item
+  goLevelDown,
+
+  /// Processing of children finished, returning to previous level
+  goLevelUp
+}
+
+/// More advanced interface than [DelegateProvider] to return information
+/// about folding tree nodes.
+abstract class BuildCoordinator {
+  ParsedItem requestData(BuildPhaseState state);
+
+  void step(BuildAction action, ParsedItem item);
+
+  static fromProvider(DelegateProvider provider) =>
+      _WrappedDelegateProvider(provider);
+}
+
+class ParsedItem {
+  final String name;
   final PNDelegate delegate;
   final dynamic data;
   final bool isLeaf;
+  final ParsedItemType type;
 
-  _ParsedItem(this.delegate, this.data, this.isLeaf);
+  /// Anything coordinator gives, it's transparent for builder, it's not later
+  /// passed to resulting processing tree.
+  final dynamic extObj;
+
+  ParsedItem.from(BuildPhaseState state, this.delegate, this.data, this.type,
+      {this.extObj})
+      : name = state.delegateName,
+        isLeaf = state.isLeaf;
+}
+
+class BuildPhaseState {
+  /// Name of xml node which children are currently processed. For root node
+  /// this is set to empty string
+  final String parentNodeName;
+
+  /// Name of xml node which is currently processed.
+  final String delegateName;
+
+  /// Current depth of const value branch. Zero means we in owner node
+  final int constValDepth;
+
+  /// True if current node has no children (end of branch)
+  final bool isLeaf;
+
+  /// Collected attributes from xml node
+  final Map<String, dynamic> data;
+
+  /// True if parsing is inside ConstValues branch of tree
+  bool get isProcessingConstValues => constValDepth > 0;
+
+  BuildPhaseState(this.parentNodeName, this.delegateName, this.constValDepth,
+      this.isLeaf, this.data);
 }
 
 class XmlTreeBuilder {
   TreeProcessor? _processor;
-  final DelegateProvider provider;
+  final BuildCoordinator coordinator;
 
-  XmlTreeBuilder(this.provider);
+  //if >0 then only constVal nodes are accepted, other will throw an exception
+  int constValDepth = 0;
+
+  XmlTreeBuilder(DelegateProvider provider)
+      : coordinator = BuildCoordinator.fromProvider(provider);
+
+  XmlTreeBuilder.coordinated(this.coordinator);
 
   /// Returns object which might be used to execute processing tree.
   /// If xml fail to parse, then [XmlParserException] is thrown. If xml is
@@ -264,46 +372,118 @@ class XmlTreeBuilder {
     return _processor!;
   }
 
+  TreeProcessor buildFrom(XmlElement xmlElement) {
+    _parseXmlRoot(xmlElement);
+    if (_processor == null) {
+      throw TreeBuilderException("Unable to build tree.");
+    }
+    return _processor!;
+  }
+
   void _parseXml(String xmlStr) {
     final xmlDoc = XmlDocument.parse(xmlStr);
-    XmlElement xmlElement = xmlDoc.rootElement;
-    final _ParsedItem item = _processElement(xmlElement);
+    _parseXmlRoot(xmlDoc.rootElement);
+  }
+
+  void _parseXmlRoot(XmlElement xmlElement) {
+    final ParsedItem item = _processElement(xmlElement);
+    if (!item.isLeaf) {
+      coordinator.step(BuildAction.goLevelDown, item);
+    }
     StackedTreeBuilder builder = StackedTreeBuilder(item.delegate, item.data);
     _processSubLevel(xmlElement, builder, false);
+    if (!item.isLeaf) {
+      coordinator.step(BuildAction.goLevelUp, item);
+    }
+    coordinator.step(BuildAction.finaliseItem, item);
     _processor = builder.build();
   }
 
   Map<String, dynamic> _extractArguments(XmlElement xmlElement) {
-    final Map<String, String> inData = {};
+    final Map<String, dynamic> inData = <String, dynamic>{};
     for (var attr in xmlElement.attributes) {
       inData[attr.name.toString()] = attr.value;
     }
     return inData;
   }
 
-  _ParsedItem _processElement(XmlElement xmlElement) {
-    final nodeName = xmlElement.name.toString();
-    final delegate = provider.delegate(nodeName);
-    final data = provider.delegateData(nodeName, _extractArguments(xmlElement));
-    if (delegate == null) {
-      throw TreeBuilderException("No delegate for xml node name '$nodeName'");
+  ParsedItem _processElement(XmlElement xmlElement) {
+    BuildPhaseState state = BuildPhaseState(
+        xmlElement.parentElement?.name.toString() ?? "",
+        xmlElement.name.toString(),
+        constValDepth,
+        xmlElement.firstElementChild == null,
+        _extractArguments(xmlElement));
+    final ParsedItem item = coordinator.requestData(state);
+    if (item.type != ParsedItemType.constValue && constValDepth > 0) {
+      throw TreeBuilderException(
+          "${item.name}: constValue node can't have non constVal nodes!");
     }
-    return _ParsedItem(delegate, data, xmlElement.firstElementChild == null);
+    coordinator.step(BuildAction.newItem, item);
+    return item;
   }
 
   void _processSubLevel(
       XmlElement xmlElement, StackedTreeBuilder builder, bool popLevelAtEnd) {
     for (var subElement in xmlElement.childElements) {
-      final _ParsedItem item = _processElement(subElement);
-      if (item.isLeaf) {
-        builder.addChild(item.delegate, item.data);
+      final ParsedItem item = _processElement(subElement);
+
+      //See documentation of ParsedItemType for details
+      if (item.type == ParsedItemType.constValue) {
+        final result = _handleAsConstValue(subElement, item);
+        if (result.value != null) {
+          builder._current.data[result.key] = result.value;
+        }
+        continue;
       } else {
-        builder.push(item.delegate, item.data);
-        _processSubLevel(subElement, builder, true);
-        if (popLevelAtEnd) {
-          builder.levelUp();
+        if (item.isLeaf) {
+          builder.addChild(item.delegate, item.data);
+          coordinator.step(BuildAction.finaliseItem, item);
+        } else {
+          builder.push(item.delegate, item.data);
+          coordinator.step(BuildAction.goLevelDown, item);
+          _processSubLevel(subElement, builder, true);
+          if (popLevelAtEnd) {
+            builder.levelUp();
+            coordinator.step(BuildAction.goLevelUp, item);
+          }
+          coordinator.step(BuildAction.finaliseItem, item);
         }
       }
     }
   }
+
+  KeyValue _handleAsConstValue(XmlElement xmlElement, ParsedItem item) {
+    constValDepth++;
+    final KeyValue result = KeyValue(item.name, null);
+    if (!item.isLeaf) {
+      coordinator.step(BuildAction.goLevelDown, item);
+
+      result.value = <String, dynamic>{};
+      for (var subElement in xmlElement.childElements) {
+        final ParsedItem subItem = _processElement(subElement);
+        final subResult = _handleAsConstValue(subElement, subItem);
+        if (subResult.value != null) {
+          result.value[subResult.key] = subResult.value;
+        }
+      }
+
+      coordinator.step(BuildAction.goLevelUp, item);
+    }
+
+    if (item.delegate(result, item.data) != Action.proceed) {
+      throw TreeBuilderException(
+          "Delegate for ${item.name} didn't return Action.proceed");
+    }
+    constValDepth--;
+    coordinator.step(BuildAction.finaliseItem, item);
+    return result;
+  }
+}
+
+class KeyValue {
+  String key;
+  dynamic value;
+
+  KeyValue(this.key, this.value);
 }