+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- Extend XmlTreeBuilder to allow coordinated build.
+- Introduce BuildCoordinator interface as an extension of DelegateProvider.
+- ConstValues node delegate get info about it parent while parsing xml.
+- ConstValues can have children.
+- Add possibility to build tree from given XmlElement rather from whole xml string.
+
 ## 0.1.1
 
 - Introduced XmlTreeBuilder with tests and example.